+++ packages/reference/src/common/EntityStore.tsx
@@ -57,7 +57,7 @@ type EntityStoreProps = {
   queryClient?: QueryClient;
 };
 
-type FetchService = ReturnType<typeof useFetch>;
+type FetchService = ReturnType<typeof _useFetch>;
 
 type GetOptions = {
   priority?: number;
@@ -218,6 +218,8 @@ type ResourceQueryKey = [
   urn: string,
   locale: string | undefined,
   referencingEntryId: string | undefined,
+  spaceId: string,
+  environmentId: string,
 ];
 
 async function fetchContentfulEntry({
@@ -357,7 +359,7 @@ async function fetchExternalResource({
   };
 }
 
-const [InternalServiceProvider, useFetch, useEntityLoader, useCurrentIds] = constate(
+const [InternalServiceProvider, _useFetch, useEntityLoader, useCurrentIds] = constate(
   function useInitServices(props: EntityStoreProps) {
     const currentSpaceId = props.sdk.ids.space;
     const currentEnvironmentId = props.sdk.ids.environmentAlias ?? props.sdk.ids.environment;
@@ -369,7 +371,7 @@ const [InternalServiceProvider, useFetch, useEntityLoader, useCurrentIds] = cons
     );
     const queryClient = useQueryClient();
     const queryCache = queryClient.getQueryCache();
-    const entityChangeUnsubscribers = useRef<Record<string, Function>>({});
+    const entityChangeUnsubscribers = useRef<Record<string, () => void>>({});
     const cmaClient = props.sdk.cma as unknown as PlainClientAPI;
     const queryQueue = useMemo(() => {
       if (props.queryConcurrency) {
@@ -556,6 +558,8 @@ const [InternalServiceProvider, useFetch, useEntityLoader, useCurrentIds] = cons
           urn,
           options?.locale,
           options?.referencingEntryId,
+          currentSpaceId,
+          currentEnvironmentId,
         ];
         return fetch(
           queryKey,
@@ -641,7 +645,7 @@ const [InternalServiceProvider, useFetch, useEntityLoader, useCurrentIds] = cons
                     return;
                   }
                   queryClient.setQueryData(query.queryKey, freshData);
-                } catch (error) {
+                } catch {
                   // If fetch fails, just invalidate the query
                   await queryClient.invalidateQueries(query.queryKey);
                 }
@@ -797,7 +801,16 @@ export function useResource<R extends Resource = Resource>(
     locale = undefined;
     referencingEntryId = undefined;
   }
-  const queryKey: ResourceQueryKey = ['Resource', resourceType, urn, locale, referencingEntryId];
+  const { space, environment } = useCurrentIds();
+  const queryKey: ResourceQueryKey = [
+    'Resource',
+    resourceType,
+    urn,
+    locale,
+    referencingEntryId,
+    space,
+    environment,
+  ];
   const { getResource } = useEntityLoader();
   const { status, data, error } = useQuery(
     queryKey,