+++ packages/rich-text/src/plugins/Hyperlink/components/LinkPopover.tsx
@@ -7,7 +7,7 @@ import { styles } from './styles';
 
 type LinkPopoverProps = {
   isLinkFocused: boolean | undefined;
-  popoverText: React.ReactNode;
+  popoverText: React.ReactNode | ((closePopover: () => void) => React.ReactNode);
   handleEditLink: () => void;
   handleRemoveLink: () => void;
   children: React.ReactNode;
@@ -26,11 +26,19 @@ export const LinkPopover = ({
 }: LinkPopoverProps) => {
   const popoverContent = React.useRef<HTMLDivElement | null>(null);
   const [isPopoverContentClicked, setIsPopoverContentClicked] = React.useState(false);
+  const isOpen = (isLinkFocused && isEditorFocused) || isPopoverContentClicked;
+  const isOpenRef = React.useRef(false);
+
+  React.useEffect(() => {
+    isOpenRef.current = isOpen;
+  }, [isOpen]);
 
   React.useEffect(() => {
     const handleMouseDown = (event) => {
       if (popoverContent.current && popoverContent.current.contains(event.target)) {
-        setIsPopoverContentClicked(true);
+        if (!isOpenRef.current) {
+          setIsPopoverContentClicked(true);
+        }
       } else {
         setIsPopoverContentClicked(false);
       }
@@ -43,25 +51,33 @@ export const LinkPopover = ({
     };
   }, []);
 
-  const isOpen = (isLinkFocused && isEditorFocused) || isPopoverContentClicked;
+  const closePopover = () => {
+    setIsPopoverContentClicked(false);
+  };
+
+  const resolvedPopoverText =
+    typeof popoverText === 'function' ? popoverText(closePopover) : popoverText;
 
   // Important to render this component in a portal
   // Otherwise the content of the popover will get copied over when users copy text from the rich text editor
 
   return (
-    // eslint-disable-next-line jsx-a11y/no-autofocus -- we don't want to autofocus the popover
     <Popover renderOnlyWhenOpen={false} usePortal={true} autoFocus={false} isOpen={isOpen}>
       <Popover.Trigger>{children}</Popover.Trigger>
       <Popover.Content className={styles.popover}>
         <Flex
           ref={popoverContent}
+          // Fix for an issue where you have to double click the copy button
+          // We intercept mousedown before Slate does, using a capture-phase listener (https://javascript.info/bubbling-and-capturing)
+          // By calling preventDefault we stop the browser from shifting focus away from the editor and the click event fires normally on the first click
+          onMouseDownCapture={(e) => e.preventDefault()}
           alignItems="center"
           paddingTop="spacing2Xs"
           paddingBottom="spacing2Xs"
           paddingRight="spacing2Xs"
           paddingLeft="spacingXs"
         >
-          {popoverText}
+          {resolvedPopoverText}
           {handleCopyLink && (
             <Tooltip placement="bottom" content="Copy link" usePortal>
               <IconButton