+++ packages/dropdown/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.4](https://github.com/contentful/field-editors/compare/@contentful/field-editor-dropdown@1.9.3...@contentful/field-editor-dropdown@1.9.4) (2026-02-23)
+
+**Note:** Version bump only for package @contentful/field-editor-dropdown
+
 ## [1.9.3](https://github.com/contentful/field-editors/compare/@contentful/field-editor-dropdown@1.9.2...@contentful/field-editor-dropdown@1.9.3) (2026-01-22)
 
 **Note:** Version bump only for package @contentful/field-editor-dropdown