+++ packages/markdown/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.15.2](https://github.com/contentful/field-editors/compare/@contentful/field-editor-markdown@1.15.1...@contentful/field-editor-markdown@1.15.2) (2026-02-23)
+
+**Note:** Version bump only for package @contentful/field-editor-markdown
+
 ## [1.15.1](https://github.com/contentful/field-editors/compare/@contentful/field-editor-markdown@1.15.0...@contentful/field-editor-markdown@1.15.1) (2026-01-23)
 
 **Note:** Version bump only for package @contentful/field-editor-markdown