+++ packages/multiple-line/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.4](https://github.com/contentful/field-editors/compare/@contentful/field-editor-multiple-line@1.7.3...@contentful/field-editor-multiple-line@1.7.4) (2026-02-23)
+
+**Note:** Version bump only for package @contentful/field-editor-multiple-line
+
 ## [1.7.3](https://github.com/contentful/field-editors/compare/@contentful/field-editor-multiple-line@1.7.2...@contentful/field-editor-multiple-line@1.7.3) (2026-01-22)
 
 **Note:** Version bump only for package @contentful/field-editor-multiple-line