+++ packages/rich-text-alpha/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.7](https://github.com/contentful/field-editors/compare/@contentful/field-editor-rich-text-alpha@0.4.6...@contentful/field-editor-rich-text-alpha@0.4.7) (2026-02-27)
+
+**Note:** Version bump only for package @contentful/field-editor-rich-text-alpha
+
+## [0.4.6](https://github.com/contentful/field-editors/compare/@contentful/field-editor-rich-text-alpha@0.4.5...@contentful/field-editor-rich-text-alpha@0.4.6) (2026-02-23)
+
+### Bug Fixes
+
+- **shared:** make React Query optional + ESM/Vite-safe via subpath export [TOL-38/46] ([#2079](https://github.com/contentful/field-editors/issues/2079)) ([64021a2](https://github.com/contentful/field-editors/commit/64021a263371c3aade8dcab93b727a012a1bce30))
+
 ## [0.4.5](https://github.com/contentful/field-editors/compare/@contentful/field-editor-rich-text-alpha@0.4.4...@contentful/field-editor-rich-text-alpha@0.4.5) (2026-01-29)
 
 **Note:** Version bump only for package @contentful/field-editor-rich-text-alpha