+++ packages/rich-text/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.20.10](https://github.com/contentful/field-editors/compare/@contentful/field-editor-rich-text@4.20.9...@contentful/field-editor-rich-text@4.20.10) (2026-02-27)
+
+**Note:** Version bump only for package @contentful/field-editor-rich-text
+
+## [4.20.9](https://github.com/contentful/field-editors/compare/@contentful/field-editor-rich-text@4.20.8...@contentful/field-editor-rich-text@4.20.9) (2026-02-23)
+
+### Bug Fixes
+
+- **shared:** make React Query optional + ESM/Vite-safe via subpath export [TOL-38/46] ([#2079](https://github.com/contentful/field-editors/issues/2079)) ([64021a2](https://github.com/contentful/field-editors/commit/64021a263371c3aade8dcab93b727a012a1bce30))
+
+## [4.20.8](https://github.com/contentful/field-editors/compare/@contentful/field-editor-rich-text@4.20.7...@contentful/field-editor-rich-text@4.20.8) (2026-02-20)
+
+### Bug Fixes
+
+- **richText:** double click issue [TOL-3510] ([#2081](https://github.com/contentful/field-editors/issues/2081)) ([97a2b51](https://github.com/contentful/field-editors/commit/97a2b51ea20f21ec258f09d2b714acb3b79e6840))
+
+## [4.20.7](https://github.com/contentful/field-editors/compare/@contentful/field-editor-rich-text@4.20.6...@contentful/field-editor-rich-text@4.20.7) (2026-02-19)
+
+### Bug Fixes
+
+- render embed block created from rich text dialog [ZEND-7619] ([#2077](https://github.com/contentful/field-editors/issues/2077)) ([249d43d](https://github.com/contentful/field-editors/commit/249d43d534dc912d60a521c454958b4f7ef81e5c))
+
+## [4.20.6](https://github.com/contentful/field-editors/compare/@contentful/field-editor-rich-text@4.20.5...@contentful/field-editor-rich-text@4.20.6) (2026-02-09)
+
+### Bug Fixes
+
+- **richText:** entity hyperlink opens slide in [TOL-3684] ([#2073](https://github.com/contentful/field-editors/issues/2073)) ([dfbbe7c](https://github.com/contentful/field-editors/commit/dfbbe7cd78be29875f01063bd24a9bbc34164f49))
+
 ## [4.20.5](https://github.com/contentful/field-editors/compare/@contentful/field-editor-rich-text@4.20.4...@contentful/field-editor-rich-text@4.20.5) (2026-01-29)
 
 **Note:** Version bump only for package @contentful/field-editor-rich-text