+++ packages/slug/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.4](https://github.com/contentful/field-editors/compare/@contentful/field-editor-slug@2.1.3...@contentful/field-editor-slug@2.1.4) (2026-02-23)
+
+**Note:** Version bump only for package @contentful/field-editor-slug
+
 ## [2.1.3](https://github.com/contentful/field-editors/compare/@contentful/field-editor-slug@2.1.2...@contentful/field-editor-slug@2.1.3) (2026-01-22)
 
 **Note:** Version bump only for package @contentful/field-editor-slug